+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- atoms-dropdown: added single-select dropdown component.
 - molecules-search-search_field: added search field form component.
 - atoms-checkbox: added checkbox form component.